+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# 1.1.1
+* Deprecated `unwrap`.
+* Service streams now pump out `e.data`, rather than the actual event.
+
 # 1.1.0+1
 * Added `unwrap`.